Rename and update model config merging functionality
The method `WriteModelConfig` in the `XmlParser` class has been renamed to `MergeModelConfig`. This new method merges a list of `PersonModel` objects (`copiedPeople`) with existing data from an XML file, updating existing entries or adding new ones as necessary. The method saves the updated list back to the XML file. The call to this method in the `InfoViewModel` class has also been updated to reflect the new name.